Improve naming in AppSec faraday patcher
