+++ lib/datadog/appsec/contrib/faraday/connection_patch.rb
@@ -9,10 +9,10 @@ module Datadog
         #
         # Wraps Faraday::Connection#initialize:
         # https://github.com/lostisland/faraday/blob/ff9dc1d1219a1bbdba95a9a4cf5d135b97247ee2/lib/faraday/connection.rb#L62-L92
-        module Connection
+        module ConnectionPatch
           def initialize(*args, &block)
             super.tap do
-              use(:datadog_appsec) unless builder.handlers.any? { |h| h.klass == Middleware }
+              use(:datadog_appsec) unless builder.handlers.any? { |h| h.klass == SSRFDetectionMiddleware }
             end
           end
         end
