+++ sig/datadog/appsec/contrib/faraday/patcher.rbs
@@ -3,8 +3,6 @@ module Datadog
     module Contrib
       module Faraday
         module Patcher
-          include Datadog::AppSec::Contrib::Patcher
-
           def self?.patched?: () -> bool
 
           def self?.target_version: () -> ::Gem::Version
