+++ sig/datadog/appsec/contrib/faraday/ssrf_detection_middleware.rbs
@@ -2,7 +2,7 @@ module Datadog
   module AppSec
     module Contrib
       module Faraday
-        class Middleware < ::Faraday::Middleware
+        class SSRFDetectionMiddleware < ::Faraday::Middleware
           def call: (untyped env) -> void
 
           private
